Fix the order of parameters in Board class tests
The height should be passed as a first parameter, and the width
as the second one. It used to be the other way, causing one of the test
to (rightly so!) fail. This commit fixes the order of both parameters
in order to reflect a recent change in Board implementation.