Reinforce the Logger utility
This commit implements several methods facilitating the use of a Logger
class. It is now easier to use the logger -, the following calls
are now available:
Log::INFO() [is a shortcut that calls Log::Logger(Log::LEVEL_INFO)]
for all the six levels of logging.

Log filtering has also been implemented. The cut-off level is held
in a structure called LoggerSettings placed in Log namespace.