Add ability to render the board at chosen place
Adds the ability to render the board at arbitrarily chosen place.
This is done by specifying the top-left pixel of the board position
as a pair of offsets, X and Y respectively.
This commit modifies the code so that the board is rendered
in the center of the game window.